+++ dart/geobase/CHANGELOG.md
@@ -14,6 +14,9 @@ NOTE: Version 0.5.0 currently [under development](https://github.com/navibyte/ge
 * [GeoJSON output for non-default coordinate reference systems #179](https://github.com/navibyte/geospatial/issues/179)
 * [WKT decoder on geobase #159](https://github.com/navibyte/geospatial/issues/159)
 
+🛠 Refactoring:
+* [Enforce "geometry" under "Feature" when encoding GeoJSON text output #91](https://github.com/navibyte/geospatial/issues/91)
+
 🐛 Bugs fixed:
 * [Clipping when projection WGS 84 to Web Mercator metric coordinates #123](https://github.com/navibyte/geospatial/issues/123)